+++ common/src/main/java/dev/ftb/mods/ftbxmodcompat/ftbquests/jei/LootCrateCategory.java
@@ -16,6 +16,7 @@ import mezz.jei.api.recipe.RecipeIngredientRole;
 import mezz.jei.api.recipe.RecipeType;
 import mezz.jei.api.recipe.category.IRecipeCategory;
 import net.minecraft.ChatFormatting;
+import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.Ingredient;
@@ -40,15 +41,17 @@ public class LootCrateCategory implements IRecipeCategory<WrappedLootCrate> {
 			}
 
 			@Override
-			public void draw(PoseStack poseStack, int xOffset, int yOffset) {
+			public void draw(GuiGraphics graphics, int xOffset, int yOffset) {
+				PoseStack poseStack = graphics.pose();
+
 				poseStack.pushPose();
 				poseStack.translate(xOffset + 8, yOffset + 8, 100);
 
 				List<WrappedLootCrate> crates = LootCrateRecipeManagerPlugin.INSTANCE.getWrappedLootCrates();
 				if (!crates.isEmpty()) {
-					GuiHelper.drawItem(poseStack, crates.get((int) ((System.currentTimeMillis() / 1000L) % crates.size())).crateStack, 0, true, null);
+					GuiHelper.drawItem(graphics, crates.get((int) ((System.currentTimeMillis() / 1000L) % crates.size())).crateStack, 0, true, null);
 				} else {
-					GuiHelper.drawItem(poseStack, new ItemStack(FTBQuestsItems.LOOTCRATE.get()), 0, true, null);
+					GuiHelper.drawItem(graphics, new ItemStack(FTBQuestsItems.LOOTCRATE.get()), 0, true, null);
 				}
 
 				poseStack.popPose();
@@ -84,10 +87,10 @@ public class LootCrateCategory implements IRecipeCategory<WrappedLootCrate> {
 					.addIngredients(() -> Ingredient.class, recipe.outputIngredients())
 					.addTooltipCallback((recipeSlotView, tooltip) -> recipeSlotView.getDisplayedIngredient()
 							.flatMap(ingr -> ingr.getIngredient(VanillaTypes.ITEM_STACK)).ifPresent(stack -> {
-								if (ItemStack.isSame(stack, recipe.outputs.get(finalSlot))) {
+								if (ItemStack.isSameItemSameTags(stack, recipe.outputs.get(finalSlot))) {
 									String chanceStr = ChatFormatting.GOLD + WeightedReward.chanceString(
-											recipe.sortedRewards.get(finalSlot).weight,
-											recipe.crate.table.getTotalWeight(true)
+											recipe.sortedRewards.get(finalSlot).getWeight(),
+											recipe.crate.getTable().getTotalWeight(true)
 									);
 									tooltip.add(Component.translatable("jei.ftbquests.lootcrates.chance", chanceStr)
 											.withStyle(ChatFormatting.GRAY));
@@ -98,7 +101,7 @@ public class LootCrateCategory implements IRecipeCategory<WrappedLootCrate> {
 	}
 
 	@Override
-	public void draw(WrappedLootCrate recipe, IRecipeSlotsView recipeSlotsView, PoseStack poseStack, double mouseX, double mouseY) {
-		LootCrateTextRenderer.drawText(poseStack, recipe.crate, 0, background.getWidth());
+	public void draw(WrappedLootCrate recipe, IRecipeSlotsView recipeSlotsView, GuiGraphics graphics, double mouseX, double mouseY) {
+		LootCrateTextRenderer.drawText(graphics, recipe.crate, 0, background.getWidth());
 	}
 }