+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0]
+
+### Added
+* Ported to Minecraft 1.20.1
+  * Functionally equivalent to last 1.19.2 release except KubeJS integration (no KubeJS for 1.20.1 at this time)
+  * Will support FTB Quests cross-mod compat when there is a public FTB Quests release for 1.20.1
+
 ## [1.2.1]
 
 ### Fixed